requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
binaryornot==0.4.4
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
debugpy==1.4.0
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.3
django-components==0.15
django-extensions==3.1.3
django-filter==2.4.0
django-ninja==0.22.2
django-tailwind==2.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
drf-access-policy==1.0.1
fastdiff==0.3.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
jedi==0.18.0
Jinja2==3.0.0
jinja2-time==0.2.0
Markdown==3.3.4
MarkupSafe==2.0.0
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
poyo==0.5.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
regex==2021.4.4
requests==2.25.1
singledispatch==3.7.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
types-requests==2.25.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
